size_t

+++ proxy/proxy.c
@@ -55,7 +55,7 @@ main(int argc, char **argv)
 	int listenfd, port;//, error;
 	int conn_to_clientfd;
 	int conn_to_serverfd;
-	int cur_bytes;	//the number of bytes read in from a single read
+	size_t cur_bytes;	//the number of bytes read in from a single read
 	unsigned int num_bytes;	//the number of bytes returned in the server response
 	
 	if (argc != 2) {
